working auth

+++ admin.php
@@ -3,16 +3,29 @@ include "init.php";
 include "dbconnect.php";
 
 use App\AdminController;
+use App\Auth;
 
 $admin_controller = new AdminController($database);
+$authPlugin = new Auth;
 
 
-switch (getenv('REQUEST_METHOD')) {
 
-    case "GET":
-        $admin_controller->show();
-        break;
-    case "POST":
-        $admin_controller->delete();
-        break;
+if ($authPlugin->is_authed()) {
+    switch (getenv('REQUEST_METHOD')) {
+        case "GET":
+            $admin_controller->show();
+            break;
+        case "POST":
+            $admin_controller->delete();
+            break;
+    }
+} else {
+    switch (getenv('REQUEST_METHOD')) {
+        case "GET":
+            $admin_controller->login_show();
+            break;
+        case "POST":
+            $admin_controller->login_post();
+            break;
+    }
 }
